% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGSimilar-package.R
\docType{package}
\name{DEGSimilar-package}
\alias{DEGSimilar}
\alias{DEGSimilar-package}
\title{DEGSimilar: A streamlined pipeline to compare DEGs via BLAST.}
\description{
DEGSimilar provides a streamlined pipeline in order to compare genes of interest
            based on BLAST databases and queries provided by the user.
}
\author{
\strong{Maintainer}: George Ly \email{george.ly@mail.utoronto.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
