% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEGSimilar.R
\name{runDEGSimilar}
\alias{runDEGSimilar}
\title{Launch Shiny App for DEGSimilar}
\usage{
runDEGSimilar()
}
\value{
No return value but opens up a Shiny page.
}
\description{
A function that launches the Shiny app for DEGSimilar.
Can perform every function of DEGSimilar, including outputting BLAST Results and graphing such results based on gene expression.
}
\examples{
\dontrun{

DEGSimilar::runDEGSimilar()
}
}
