% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGSimilar-package.R
\docType{package}
\name{DEGSimilar-package}
\alias{DEGSimilar}
\alias{DEGSimilar-package}
\title{DEGSimilar: A streamlined pipeline to compare DEGs via BLAST.}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: George Ly \email{george.ly@mail.utoronto.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
